#75 Simplified project layout and moved code under pkg. Created separate packages for each secret provider and split up apis and controller packages.
