+++ deployments/helm/templates/role_binding_watch.yaml
