+++ pkg/onepassword/op.go
